requirements.txt
openai
python-dotenv
pydantic 